requirements.txt
efficientnet_pytorch==0.7.1
matplotlib==3.8.4
numpy==2.2.1
PyYAML==6.0.2
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.4
